pyproject.toml
[tool.poetry]
name = "local-airflow"
version = "0.1.0"
description = ""
authors = ["Ryoto Ogawa <71643807+ryotoitoi@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "local_airflow"}]

[tool.poetry.dependencies]
python = "^3.9"
apache-airflow = "^2.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
